requirements.txt
asgiref==3.6.0
certifi==2022.5.18.1
distlib==0.3.4
Django==4.1.6
django-filter==22.1
djangorestframework==3.14.0
filelock==3.7.1
Markdown==3.4.1
mysql-connector-python==8.0.29
Pillow==9.4.0
pipenv==2022.6.7
platformdirs==2.5.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.14.1
virtualenv-clone==0.5.7
